requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
axios==0.4.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses-json==0.6.7
Deprecated==1.2.14
diffusers==0.29.1
distro==1.9.0
dnspython==2.6.1
efficientnet==1.1.1
efficientnet_pytorch==0.7.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.3
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-trans-new==1.1.9
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==1.*
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Keras-Applications==1.0.8
kiwisolver==1.4.5
ko-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.7.0/ko_core_news_sm-3.7.0-py3-none-any.whl#sha256=b1a15a4987a8f9835031a6bd2fe57fe158097ab5304221c41df1bd4aab8cf458
kubernetes==30.1.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-experimental==0.0.63
langchain-huggingface==0.0.3
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.82
language_data==1.2.0
lazy_loader==0.4
libretranslatepy==2.1.1
lmdb==1.5.1
loguru==0.7.2
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
#openai==0.28.0
#openai==1.32.0
openai>=1.32.0,<2.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
paddleocr==2.8.1
pandas==2.2.2
pillow==10.3.0
posthog==3.5.0
preshed==3.0.9
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
router==0.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.0.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.37.2
streamlit==1.36.0
sympy==1.12.1
tbb==2021.12.0
tenacity==8.4.2
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.7.2
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
transforms==0.2.1
translate==3.6.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.22.0
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2